\name{ecoregions}


\alias{ecoregions}
 
\title{
Global ecoregions polygons data
}

\description{
Get global ecoregions data.
The data are from Olson et al (WWF) as modified by TNC. Downloaded from https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/WTLNRG
}

\usage{
ecoregions()
}

\value{
SpatVector with ecoregion polygons
}

\examples{
e <- ecoregions()
}




